Move rich detecting and formatting to utils.py
