+++ kedro/logging.py
@@ -1,6 +1,5 @@
 import logging
 import sys
-from functools import lru_cache
 from pathlib import Path
 from typing import Any
 
@@ -55,14 +54,3 @@ class RichHandler(rich.logging.RichHandler):
             # fixed on their side at some point, but until then we disable it.
             # See https://github.com/Textualize/rich/issues/2455
             rich.traceback.install(**traceback_install_kwargs)  # type: ignore[arg-type]
-
-
-@lru_cache(maxsize=None)
-def _has_rich_handler(logger: logging.Logger) -> bool:
-    """Returns true if the logger has a RichHandler attached."""
-    return any(isinstance(handler, RichHandler) for handler in logger.handlers)
-
-
-def _format_rich(value: str, markup: str) -> str:
-    """Format string with rich markup"""
-    return f"[{markup}]{value}[/{markup}]"
